Cap utilization at 1 for the supply rate scalar
This makes the supply rate linear after util = 1, which guarantees it won't surpass borrow rate and eat into reserves, or eat into reserve rate at all, which the quadratic form does.

+++ contracts/CometMath.sol
@@ -16,7 +16,7 @@ contract CometMath {
     error InvalidInt256();
     error NegativeNumber();
 
-    function min(uint104 a, uint104 b) internal pure returns (uint104) {
+    function min(uint a, uint b) internal pure returns (uint) {
         return a < b ? a : b;
     }
